Fixed issue with title score not playing

+++ arduboy-game/MusicPlayer.h
@@ -6,7 +6,7 @@
 #include <Arduino.h>
 #include <avr/pgmspace.h>
 
-extern void playMusic(int song);
+extern void playMusic(byte song);
 extern void stopMusic();
 
 extern void sfx(byte tone);
